Expand morph range
Now, if morph CV is maxed at 5V, turning the morph knob to the right will continue morphing past the algorithm on the right and onward toward the algorithm on the left.

This works very well together with cicular morphing (eg. with ZZC Clock or Rampage). Now the knob doesn't break the circular morphing.